+++ code/input.txt
@@ -1,4 +1,4 @@
 capacitor John is 15 C
-resistor Harry is 15 Om
+resistor Harry is 15 m
 connect John and Harry in series
-
+#automatically print out the result
